move achievement data to its own file
